service/Service/requirements.txt
fastapi==0.68.0
sqlalchemy==1.4.23
uvicorn==0.15.0
pydantic==1.8.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.5
werkzeug==2.0.1
email-validator==1.1.3
python-dotenv==0.19.0
redis==4.0.2
phonenumbers==8.12.33 